Remove joystick-related functionality and event waiting
This commit removes all the joystick-related functionality that was present in the glfw library. Additionally, the functions to wait for events have also been removed. This includes the removal of the connection and disconnection events for the joystick. All the related code for these functionalities, including function declarations and definitions, has been removed.