Remove obsolete GLFW documentation files
This revision eliminates various obsolete GLFW files. These files contained documentation content regarding engine and graphic libraries, which are no longer used in the current framework. Their removal should result in a simpler, cleaner codebase. It should be noted that this is purely a code maintenance task and it is not expected to affect the functionality or performance of the application.