Remove OpenGL and OpenGL ES headers from glfw3.h
The commit removes the inclusion of OpenGL and OpenGL ES headers from glfw3.h file in the GLFW library. It also refactors the test examples to adjust to this change by altering the way libraries are linked and adjustments in the title.cpp file. This clean-up helps in removing unnecessary dependencies and potential conflicts with different OpenGL versions or extensions.