Relocate DEPTH_TEST definition and add Pipeline::enable method
In this change, the GL_DEPTH_TEST definition was moved from the glad.h file to the macros.hpp file under the OpenGL graphics engine's code directory. Additionally, an enable method was added to the Pipeline class within the pipeline.cpp file to enable flags in OpenGL. This commit helps to better organize our OpenGL definitions and enhance functionality.