Remove swap interval functionality and `title.c` test file
The swap interval functionality is removed from glfw including relevant changes in `glfw3.h` and `context.c`. Also, a test file `title.c` from the `tests` directory that was related to setting swap interval is deleted. Changes in the `CMakeLists.txt` reflect the removal of `title.c`. References to `glfwSwapInterval` in `triangle-opengl.c` have been eradicated as part of this update.