Remove input drop and joystick functionalities from GLFW
This commit removes the joystick test source file, the drag and drop input mode, and all related functionalities from GLFW. This includes changes in the CMakeLists files, engine files, test files, and several internal files in the GLFW library. The update streamlines the library and focuses on core functionalities instead.