Remove window position callback functionality
The callback functionality for window position in the GLFW library has been removed. The deleted code includes the function definitions, function calls, and related comments. Users will no longer be able to set callbacks arising from changes to the position of the window as this feature is no longer supported.