Refactor engine code and modify third-party GLFW library
In this commit, we have refactored several parts of our platform-independent engine code. In the `base` and `win32` directories, we've updated the Window class to include new methods and handle its own platform-specific window creation, destruction, and updating. We also introduced the WindowConfig and WindowSize configurations that will be passed into the `create` method. More platform settings were updated in the corresponding `public.hpp` file.
Additionally, we customized the GLFW library, simplifying certain functionalities and removing unnecessary services or features not needed in our project, such as the WindowClose and WindowSize features, along with others in the internal and win32 parts of GLFW. Please be aware that these changes make it incompatible with the official GLFW.