Refactor CMakeLists and context code in glfw library
The commit simplifies the CMakeLists.txt file in the glfw library and refactors the code in the context.c file. In the CMakeLists.txt, unnecessary options and settings were removed for a cleaner configuration while the context.c file saw a reduction in redundant variable declarations and simpler loop iterations. Moreover, library and